Cleanup

+++ Mima-Sim/MIMA/Instructions/Arithmetik/IncReg.cs
@@ -12,7 +12,7 @@ namespace MimaSim.MIMA.Instructions.Arithmetik
             var oldValue = cpu.GetRegister(Registers.Accumulator);
             var newValue = oldValue + 1;
 
-            cpu.SetRegister(Registers.Accumulator, (byte)newValue);
+            cpu.SetRegister(Registers.Accumulator, (short)newValue);
 
             return false;
         }
